+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
-## 1.3.0
+## 1.X.X
 
 `2021-05-XX`
 
 ### What's new
 
+- **Added new `FileUpload` component** with drag & drop functionality.
 - **Added lazy loading to `Table`** with new `onBottomReached` and `onRowInViewport` props.
 
 ### Fixes